Tools: Initial commit of Redfish Profile Simulator
This is an open source project on DMTF GitHub.
(https://github.com/DMTF/Redfish-Profile-Simulator)
This tool simulates the HTTP request methods (POST, PATCH, PUT, GET) on
Redfish resource maintained by Redfish Profile Simulator.
EDK2 open source uses this simulator for the use case when Redfish service
has not been set up on the platform, or for the quick Redfish firmware
feature development.

We clone this project under edk2-redfish-client and maintain because
this project has currently been using and updating rarely.
That is easier for edk2-redfish-client to add features to the simulator
or modify the simulator to align with edk2 requirement on Redfish service.

The license of this tool is on the term of BSD 3-Clause License.
Refer to LICENSE.md.

Signed-off-by: Nickle Wang <nicklew@nvidia.com>
Cc: Abner Chang <abner.chang@amd.com>
Cc: Igor Kulchytskyy <igork@ami.com>
Reviewed-by: Abner Chang <abner.chang@amd.com>