RedfishClientPkg: Add EDK2 Redfish JSON C Structure stuff into build
Add edk2 Redfish JSON to C Structure libraries into build.
Add edk2 Redfish JSON to C Structure DXE drivers into build.
REDFISH_CLIENT_ALL_AUTOGENED macro is introduced to skip building
those edk2 Redfish JSON to C Structure libraries/drivers because
it is too many.

This commit adds below Redfish JSON C Structure stuff into the build,
- edk2 Redfish JSON to C Structure libraries
- edk2 Redfish JSON to C Structure DXE drivers

There are three relevant commits to this one those were not sent
for the review because those are script auto-generated files and
too many. (+18000 files).

1. Redfish JSON schema to C structure source files
2. RedfishClientPkg/Converter: EDK2 Redfish JSON schema to C
   structure DXE drivers
3. EDK2 Redfish JSON schema to C structure libraries

Above drivers/librares are not all required to be built, only the
certain instances need to be pull in the build process according to the platform demands.

On edk2 Redfish client implementation, only the Redfish schema
which referred in the HII form requires the corresponding EDK2
Redfish JSON C Structure drivers/libraries to build. This is
described in the item #10 in EDK2 Redfish Client implementation
guide at the below link.
https://github.com/tianocore/edk2-redfish-client/blob/main/RedfishClientPkg/Readme.md

We also propose to create another GitHub repo under Tianocore for
the entire RedfishClientPkg. With this we won't add overheads to
the regular edk2 repository because the tons of files.

Signed-off-by: Abner Chang <abner.chang@amd.com>
Cc: Nickle Wang <nicklew@nvidia.com>
Cc: Igor Kulchytskyy <igork@ami.com>
Reviewed-by: Nickle Wang <nicklew@nvidia.com>