RedfishClientPkg/RedfishFeatureUtilityLib: fix issues and enhancement
-Add RedfishDebugLib to print Redfish request and response
details when PATCH and POST request return error.
-Use "%a:" in debug print to align with the style in EDK2.
-Enhance GetConfigureLang() to handle pending resource URI.
Pending resource share ths same schema as its active resource.
So we can simply remove "/Settings" from URI and search again.
-Enhancement to GetPropertyVagueValue(). This function stops
to handle remaining property while error happens. We like to
process as much as properties as we can and skip the problematic
property.
-Enhancement to MatchPropertyWithJsonContext(). For string type
of BIOS attribute, it is possible that the attribute value is
empty by default. This does not means that the resource is not
provisioned yet.
-Fix typo.

Signed-off-by: Nickle Wang <nicklew@nvidia.com>
Cc: Abner Chang <abner.chang@amd.com>
Cc: Igor Kulchytskyy <igork@ami.com>
Cc: Nick Ramirez <nramirez@nvidia.com>
Reviewed-by: Abner Chang <abner.chang@amd.com>