RedfishClientPkg/RedfishFeatureCoreDxe: report failure via status code
Manageability status code is introduced to edk2. Add the ability
to report Redfish communication failure via status code. This
gives the chance for BMC to capture Redfish error.

When system reboot is required to apply Redfish changes, we also report
it via status code. So, user is aware of this reboot.

Signed-off-by: Nickle Wang <nicklew@nvidia.com>